refactor(接口测试): 接口测试执行相关代码整理优化
