+++ backend/src/main/java/io/metersphere/api/exec/queue/DBTestQueue.java
@@ -10,6 +10,6 @@ import java.util.Map;
 @Data
 public class DBTestQueue extends ApiExecutionQueue {
     private String completedReportId;
-    private ApiExecutionQueueDetail queue;
+    private ApiExecutionQueueDetail detail;
     private Map<String, String> detailMap = new HashMap<>();
 }
